+++ routes/auth.js
@@ -2,8 +2,8 @@ const express = require('express');
 
 const router = express.Router();
 
-const { signUpUser } = require('../controllers/auth');
-const { loggIn } = require('../controllers/auth');
+const { signUpUser } = require('../controllers/authController');
+const { loggIn } = require('../controllers/authController');
 
 router.post('/signup', signUpUser);
